requirements.txt
Flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flask-jwt-extended==4.5.3
flask-swagger-ui==4.11.1
flask-cors==4.0.0
PyJWT==2.8.0
Flask-Mail==0.9.1 